Fix GCC atomic_ref link error
Depending on how GCC was built, it might need "-latomic"
to link C++ programs that use std::atomic_ref.
This commit adds a LINALG_ENABLE_ATOMIC_REF CMake option
that is OFF by default.
I've tested the option with my GCC 9.4 build,
which happens not to need "-latomic";
the linker succeeds even when including that flag.