Fix triangular_matrix_matrix_solve
The sequential implementation of triangular_matrix_matrix_solve
for the lower triangular, right side case had a j loop that was
counting in the wrong direction: up, instead of down.

This commit fixes that, and adds a regression test.
The test also uncovered a compilation bug that this commit fixes.

Fixes #244.