Fix layout_transpose
layout_transpose was imitating an old version of mdspan.
For example, it used size_type instead of index_type
as the return type of operator().  That made tests for
the return type of transposed fail (because the transposed
layout mapping's extents_type had the wrong index_type).
This commit fixes that.