Rename accessor_conjugate to conjugated_accessor
and otherwise bring it in line with the Standard.

Add conjugated test for a nonarithmetic real type.
"Real type" in this case just means that conj of it
is not ADL-findable.

Add conjugated test for a custom complex number.
"Custom complex number" in this case just means
that conj of it is ADL-findable.